Cambio el tipo del id de las entidades de Long a Integer porque la base no soporta el Long.

+++ gestionTurnos/src/main/java/ar/com/trivoli/gestionturnos/modeloturno/model/ModeloTurno.java
@@ -17,7 +17,7 @@ import ar.com.trivoli.gestionturnos.common.model.EntidadBase;
  */
 @Entity
 @Table(name = "templateReservas")
-public class ModeloTurno extends EntidadBase<Long> {
+public class ModeloTurno extends EntidadBase<Integer> {
 	private String descripcion;
 
 	@Column(name = "titulo")
